servicio de generacion de latex y pdf

+++ app/Http/Controllers/ExamLayoutController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\ExamStatusEnum;
 use App\Models\Exam;
 use App\Models\ExamLayout;
 use Illuminate\Http\Request;
 
-class ExamController extends Controller
+class ExamLayoutController extends Controller
 {
     /**
      * Display the specified resource.
@@ -27,7 +28,17 @@ class ExamController extends Controller
      */
     public function destroy(Exam $exam)
     {
+        if($exam->status !== ExamStatusEnum::VARIATED){
+            return response()->json([
+                'success' => false,
+                'message' => 'El examen debe estar en estado PERMUTADO para eliminar los temas.'
+            ], 400);
+        }
+
         ExamLayout::where('exam_id', $exam->id)->delete();
+        $exam->status = ExamStatusEnum::MASTERED;
+        $exam->save();
+        
         return response()->json(null, 204);
     }
 }
